requirements.txt
playwright==1.42.0
aiohttp==3.9.3
python-dotenv==1.0.1